backend/requirements.txt
flask 
flask-cors 
flask-sqlalchemy 
flask-bcrypt 
flask-jwt-extended
weasyprint
python-dotenv
psycopg2
gunicorn==19.7.1
requests
redis
flask-migrate
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.1
